Refactor benchmarks with clearer naming and organization
Rename benchmarks from "qN_" research question format to descriptive
"X_vs_Y" naming that communicates what's being compared:

- u8_vs_u64: Byte iteration vs u64 chunk processing
- array_vs_slice: Fixed arrays vs dynamic slices
- batch_vs_single: Batch API vs loop of single calls
- batch_inlined_vs_call: Inlining strategies in batch loops
- multiversion: Static compilation vs runtime CPU dispatch (x86 only)
- batch_multiversion: Multiversion inlining strategies (x86 only)
- vs_competitors: This crate vs other Hamming distance libraries

Key changes:
- Benchmarks use inlined implementations instead of public API
  (except vs_competitors) for stability against API changes
- x86-only benchmarks show helpful messages on other platforms
- multiversion benchmark compares static vs runtime dispatch
  side-by-side in a single run
