+++ benches/u8_vs_u64.rs
@@ -1,10 +1,10 @@
-//! Q1: What's the fastest way to compute Hamming distance on u8 arrays?
+//! Compares byte-by-byte iteration (u8) vs u64 chunk processing.
 //!
-//! Key questions:
-//! - Is byte-by-byte iteration faster or slower than chunks_exact(8)?
-//! - Does the remainder handling in chunks_exact add overhead?
+//! Key insight: On ARM, simple u8 iteration auto-vectorizes well with NEON.
+//! On x86, u64 chunk processing enables AVX-512 VPOPCNTDQ when available.
 //!
-//! Run with: cargo bench --bench q1_data_types
+//! Run with: cargo bench --bench u8_vs_u64
+//! Compare with: RUSTFLAGS="-C target-cpu=native" cargo bench --bench u8_vs_u64
 
 mod helpers;
 
@@ -15,10 +15,10 @@ fn main() {
 }
 
 // ============================================================================
-// Implementations to benchmark
+// Implementations
 // ============================================================================
 
-/// Byte-by-byte iteration.
+/// Byte-by-byte iteration - simple, auto-vectorizes well on ARM.
 #[inline]
 fn hamming_u8_iter<const N: usize>(a: &[u8; N], b: &[u8; N]) -> u32 {
     a.iter()
@@ -27,22 +27,9 @@ fn hamming_u8_iter<const N: usize>(a: &[u8; N], b: &[u8; N]) -> u32 {
         .sum()
 }
 
-/// Process as u64 chunks (no remainder handling).
+/// Process as u64 chunks - enables AVX-512 VPOPCNTDQ on x86.
 #[inline]
-fn hamming_u8_chunks<const N: usize>(a: &[u8; N], b: &[u8; N]) -> u32 {
-    a.chunks_exact(8)
-        .zip(b.chunks_exact(8))
-        .map(|(a_chunk, b_chunk)| {
-            let a_val = u64::from_ne_bytes(a_chunk.try_into().unwrap());
-            let b_val = u64::from_ne_bytes(b_chunk.try_into().unwrap());
-            (a_val ^ b_val).count_ones()
-        })
-        .sum()
-}
-
-/// Process as u64 chunks with byte-by-byte remainder handling.
-#[inline]
-fn hamming_u8_chunks_with_remainder<const N: usize>(a: &[u8; N], b: &[u8; N]) -> u32 {
+fn hamming_u64_chunks<const N: usize>(a: &[u8; N], b: &[u8; N]) -> u32 {
     let a_chunks = a.chunks_exact(8);
     let b_chunks = b.chunks_exact(8);
 
@@ -67,7 +54,7 @@ fn hamming_u8_chunks_with_remainder<const N: usize>(a: &[u8; N], b: &[u8; N]) ->
 }
 
 // ============================================================================
-// Benchmarks
+// Benchmarks (sizes in bits: 512b, 768b, 1024b, 2048b = 64, 96, 128, 256 bytes)
 // ============================================================================
 
 #[divan::bench(consts = [64, 96, 128, 256])]
@@ -78,15 +65,8 @@ fn u8_iter<const N: usize>(bencher: divan::Bencher) {
 }
 
 #[divan::bench(consts = [64, 96, 128, 256])]
-fn u8_chunks<const N: usize>(bencher: divan::Bencher) {
+fn u64_chunks<const N: usize>(bencher: divan::Bencher) {
     let a: [u8; N] = random_bytes();
     let b: [u8; N] = random_bytes();
-    bencher.bench_local(|| hamming_u8_chunks(&a, &b));
-}
-
-#[divan::bench(consts = [64, 96, 128, 256])]
-fn u8_chunks_with_remainder<const N: usize>(bencher: divan::Bencher) {
-    let a: [u8; N] = random_bytes();
-    let b: [u8; N] = random_bytes();
-    bencher.bench_local(|| hamming_u8_chunks_with_remainder(&a, &b));
+    bencher.bench_local(|| hamming_u64_chunks(&a, &b));
 }
